+++ src/cli/commands/public-cmds/isolate-cmd.js
@@ -11,6 +11,8 @@ export default class Isolate extends Command {
     ['w', 'write-bit-dependencies [boolean] ', 'write bit components dependencies to package.json file'],
     ['l', 'npm-links [boolean]', 'point dependencies link files to npm package'],
     ['i', 'install-packages [boolean]', 'install npm packaged dependencies'],
+    ['', 'dist', 'write dist files (when exist) to the configured directory'],
+    ['', 'conf', 'write the configuration file (bit.json)'],
     ['', 'no-package-json [boolean]', 'do not generate package.json for the isolated component'],
     ['o', 'override [boolean]', 'override existing isolated component']
   ];
@@ -23,10 +25,14 @@ export default class Isolate extends Command {
       writeBitDependencies: ?boolean,
       npmLinks: ?boolean,
       installPackages: ?boolean,
+      dist: ?boolean,
+      conf: ?boolean,
       noPackageJson: ?boolean,
       override: ?boolean
     }
   ): Promise<any> {
+    opts.writeToPath = opts.directory;
+    delete opts.directory;
     return isolate(id, scopePath || process.cwd(), opts);
   }