+++ src/consumer/consumer.js
@@ -203,7 +203,7 @@ export default class Consumer {
       if (componentMap.rootDir) {
         bitDir = path.join(bitDir, componentMap.rootDir);
       }
-      const componentModule = await this.scope.sources.get(id);
+      const componentFromModel = await this.scope.getFromLocalIfExist(idWithConcreteVersion);
       let component;
       try {
         component = await Component.loadFromFileSystem({
@@ -213,7 +213,7 @@ export default class Consumer {
           id: idWithConcreteVersion,
           consumerPath: this.getPath(),
           bitMap,
-          deprecated: componentModule ? componentModule.deprecated : false
+          componentFromModel
         });
       } catch (err) {
         if (throwOnFailure) throw err;
@@ -262,7 +262,9 @@ export default class Consumer {
     withEnvironments: ?boolean,
     cache?: boolean = true,
     withPackageJson?: boolean = true,
-    force?: boolean = false
+    force?: boolean = false,
+    dist?: boolean = false,
+    conf?: boolean = false
   ): Promise<> {
     const dependenciesFromBitJson = BitIds.fromObject(this.bitJson.dependencies);
     const bitMap = await this.getBitMap();
@@ -284,13 +286,24 @@ export default class Consumer {
     let componentsAndDependenciesBitMap = [];
     if (dependenciesFromBitJson) {
       componentsAndDependenciesBitJson = await this.scope.getManyWithAllVersions(dependenciesFromBitJson, cache);
-      await this.writeToComponentsDir(componentsAndDependenciesBitJson, undefined, undefined, withPackageJson);
+      await this.writeToComponentsDir({
+        componentsWithDependencies: componentsAndDependenciesBitJson,
+        withPackageJson,
+        withBitJson: conf,
+        dist
+      });
     }
     if (componentsFromBitMap.length) {
       componentsAndDependenciesBitMap = await this.scope.getManyWithAllVersions(componentsFromBitMap, cache);
-      // Don't write the package.json for an authored component, because it's dependencies probably managed
-      // By the root packge.json
-      await this.writeToComponentsDir(componentsAndDependenciesBitMap, undefined, false, false);
+      // Don't write the package.json for an authored component, because its dependencies probably managed by the root
+      // package.json
+      await this.writeToComponentsDir({
+        componentsWithDependencies: componentsAndDependenciesBitMap,
+        force: false,
+        withPackageJson: false,
+        withBitJson: conf,
+        dist
+      });
     }
     const componentsAndDependencies = [...componentsAndDependenciesBitJson, ...componentsAndDependenciesBitMap];
     if (withEnvironments) {
@@ -327,7 +340,9 @@ export default class Consumer {
     writeToPath?: string,
     withPackageJson?: boolean = true,
     writeBitDependencies?: boolean = false,
-    force?: boolean
+    force?: boolean = false,
+    dist?: boolean = false,
+    conf?: boolean = false
   ) {
     logger.debug(`importSpecificComponents, Ids: ${rawIds.join(', ')}`);
     if (!force) {
@@ -335,15 +350,16 @@ export default class Consumer {
     }
     // $FlowFixMe - we check if there are bitIds before we call this function
     const bitIds = rawIds.map(raw => BitId.parse(raw));
-    const componentDependenciesArr = await this.scope.getManyWithAllVersions(bitIds, cache);
-    await this.writeToComponentsDir(
-      componentDependenciesArr,
+    const componentsWithDependencies = await this.scope.getManyWithAllVersions(bitIds, cache);
+    await this.writeToComponentsDir({
+      componentsWithDependencies,
       writeToPath,
-      undefined,
       withPackageJson,
-      writeBitDependencies
-    );
-    return { dependencies: componentDependenciesArr };
+      withBitJson: conf,
+      writeBitDependencies,
+      dist
+    });
+    return { dependencies: componentsWithDependencies };
   }
 
   import(
@@ -354,13 +370,32 @@ export default class Consumer {
     writeToPath?: string,
     withPackageJson?: boolean = true,
     writeBitDependencies?: boolean = false,
-    force?: boolean = false
+    force?: boolean = false,
+    dist?: boolean = false,
+    conf?: boolean = false
   ): Promise<{ dependencies: ComponentWithDependencies[], envDependencies?: Component[] }> {
     loader.start(BEFORE_IMPORT_ACTION);
     if (!rawIds || R.isEmpty(rawIds)) {
-      return this.importAccordingToBitJsonAndBitMap(verbose, withEnvironments, cache, withPackageJson, force);
+      return this.importAccordingToBitJsonAndBitMap(
+        verbose,
+        withEnvironments,
+        cache,
+        withPackageJson,
+        force,
+        dist,
+        conf
+      );
     }
-    return this.importSpecificComponents(rawIds, cache, writeToPath, withPackageJson, writeBitDependencies, force);
+    return this.importSpecificComponents(
+      rawIds,
+      cache,
+      writeToPath,
+      withPackageJson,
+      writeBitDependencies,
+      force,
+      dist,
+      conf
+    );
   }
 
   importEnvironment(rawId: ?string, verbose?: boolean) {
@@ -398,14 +433,25 @@ export default class Consumer {
    * In case there are some same dependencies shared between the components, it makes sure to
    * write them only once.
    */
-  async writeToComponentsDir(
+  async writeToComponentsDir({
+    componentsWithDependencies,
+    writeToPath,
+    force = true,
+    withPackageJson = true,
+    withBitJson = true,
+    writeBitDependencies = false,
+    createNpmLinkFiles = false,
+    dist = true
+  }: {
     componentsWithDependencies: ComponentWithDependencies[],
     writeToPath?: string,
-    force?: boolean = true,
-    withPackageJson?: boolean = true,
-    writeBitDependencies?: boolean = false,
-    createNpmLinkFiles?: boolean = false
-  ): Promise<Component[]> {
+    force?: boolean,
+    withPackageJson?: boolean,
+    withBitJson?: boolean,
+    writeBitDependencies?: boolean,
+    createNpmLinkFiles?: boolean,
+    dist?: boolean
+  }): Promise<Component[]> {
     const bitMap: BitMap = await this.getBitMap();
     const dependenciesIdsCache = [];
     const writeComponentsP = componentsWithDependencies.map((componentWithDeps) => {
@@ -418,16 +464,25 @@ export default class Consumer {
         componentMap && componentMap.origin === COMPONENT_ORIGINS.AUTHORED
           ? COMPONENT_ORIGINS.AUTHORED
           : COMPONENT_ORIGINS.IMPORTED;
-      if (origin === COMPONENT_ORIGINS.IMPORTED) componentWithDeps.component.stripOriginallySharedDir(bitMap);
+      if (origin === COMPONENT_ORIGINS.IMPORTED) {
+        componentWithDeps.component.stripOriginallySharedDir(bitMap);
+        componentWithDeps.component.setDistDir(this.bitJson.distTarget);
+        componentWithDeps.component.updateDistsLocation();
+      }
+      // don't write dists files for authored components as the author has its own mechanism to generate them
+      // also, don't write dists file for imported component, unless the user used '--dist' flag
+      const writeDistsFiles = dist && origin === COMPONENT_ORIGINS.IMPORTED;
       return componentWithDeps.component.write({
         bitDir,
         force,
         bitMap,
+        withBitJson,
         withPackageJson,
         origin,
         consumerPath: this.getPath(),
         driver: this.driver,
         writeBitDependencies,
+        writeDistsFiles,
         dependencies: componentWithDeps.dependencies,
         componentMap
       });